First commit: converting Matlab types into Python types
NEW FILES:
    matarray2numpyarry.m - converts a Matlab array into a Python Numpy array.
Gives the option to match the dimension ordering so that (if M is the Matlab
array and P the Python one) M(1,:,:) == P[0,:,:]. Alternately, the native
ordering can be maintained, so that M(1,:,:) == P[:,:,0].

    struct2pydict.m - converts a scalar Matlab struct into a Python dictionary,
or a multi-element struct into a list of dictionaries. Currently does not handle
fields that are arrays or structures well.